+++ src/router/index.ts
@@ -12,11 +12,11 @@ import ReturnDetails from '@/views/ReturnDetails.vue'
 import TransferOrders from '@/views/TransferOrders.vue';
 import TransferOrderDetail from '@/views/TransferOrderDetail.vue';
 import Notifications from '@/views/Notifications.vue'
-import Login from '@/views/Login.vue'
+import Login from '@common/components/Login.vue'
 
 import { useUserStore } from '@/store/user';
 import { translate, commonUtil } from '@common';
-import { useAuth } from '@/composables/useAuth'
+import { useAuth } from '@common/composables/auth'
 
 import { businessOutline, calendar, download, gitPullRequestOutline, settingsOutline } from "ionicons/icons";
